emacs: Don't cleanup whitespaces in org-mode
This must be the 5th time I have to fight with git because I forgot to
disable whitespace-cleanup at the start of my org session and ended up
butchering the indentation of all my python blocks.

Eventually should hack something up to do the right thing depending on
the context of the block being cleaned, but this is much simpler - and
whitespace-cleanup should still work on snippet editing.